Add support for building + testing both C++ & C# with CMake.
This commit adds the ability to build both codebases with CMake. This is
of course primarily aimed at portability for building with C++, but is
nonetheless quite convenient for working on both of them at the same
time.

C# compilation is disabled by default since it requires you to be using
the Visual Studio generator in order to work - however, you should be
able to invoke the resulting project files from MSBuild if you don't
have or want the full Visual Studio IDE.

The code of the applications has been altered to separate unit tests
out. They now do not run unless there is some explicit call to invoke
them. This ensures that when CTest runs, it is able to execute the
different kinds of test separately, preventing mixing of results in the
event of a failure.